+++ KaguyaProjectV2/KaguyaBot/Core/Commands/Utility/MyReminders.cs
@@ -8,6 +8,7 @@ using KaguyaProjectV2.KaguyaBot.Core.KaguyaEmbed;
 using KaguyaProjectV2.KaguyaBot.DataStorage.DbData.Queries;
 using System;
 using System.Linq;
+using System.Text;
 using System.Threading.Tasks;
 
 namespace KaguyaProjectV2.KaguyaBot.Core.Commands.Utility
@@ -23,19 +24,45 @@ namespace KaguyaProjectV2.KaguyaBot.Core.Commands.Utility
         public async Task Command()
         {
             var user = await DatabaseQueries.GetOrCreateUserAsync(Context.User.Id);
-            var reminders = user.Reminders.ToList();
+            var reminders = user.Reminders.Where(x => x.Expiration > DateTime.Now.ToOADate()).ToArray();
             var embed = new KaguyaEmbedBuilder();
 
             int i = 0;
-            foreach (var reminder in reminders)
+
+            if(!(reminders.Length == 0))
             {
-                i++;
+                foreach (var reminder in reminders)
+                {
+                    i++;
+
+                    var expirationStr = DateTime.FromOADate(reminder.Expiration).Humanize(false);
+
+                    var fSb = new StringBuilder();
+                    fSb.AppendLine($"Reminder: `{reminder.Text}`");
+                    fSb.AppendLine($"Expires: `{expirationStr}`");
+                    
+                    var field = new EmbedFieldBuilder
+                    {
+                        IsInline = false,
+                        Name = $"#{i}",
+                        Value = fSb.ToString()
+                    };
+
+                    embed.AddField(field);
+                }
+
+                embed.Footer = new EmbedFooterBuilder
+                {
+                    Text = "To delete a reminder, click the corresponding reaction."
+                };
+            }
+
+            else
+            {
                 var field = new EmbedFieldBuilder
                 {
-                    IsInline = true,
-                    Name = $"#{i}",
-                    Value = $"Reminder: `{reminder.Text}`\n" +
-                            $"Notification: `{DateTime.FromOADate(reminder.Expiration).Humanize(false)}`"
+                    Name = "No reminders active",
+                    Value = "You currently don't have any active reminders."
                 };
 
                 embed.AddField(field);
@@ -45,7 +72,7 @@ namespace KaguyaProjectV2.KaguyaBot.Core.Commands.Utility
             var data = new ReactionCallbackData("", embed.Build());
             foreach (var reminder in reminders)
             {
-                var callback = new ReactionCallbackItem(GlobalProperties.EmojisOneThroughNine()[j], 
+                data.AddCallBack(GlobalProperties.EmojisOneThroughNine()[j], 
                 async (c, r) =>
                 {
                     await DatabaseQueries.DeleteAsync(reminder);
@@ -54,6 +81,8 @@ namespace KaguyaProjectV2.KaguyaBot.Core.Commands.Utility
 
                 j++;
             }
+
+            await InlineReactionReplyAsync(data);
         }
     }
 }