+++ KaguyaProjectV2/KaguyaBot/Core/Commands/Utility/RemindMe.cs
@@ -24,8 +24,8 @@ namespace KaguyaProjectV2.KaguyaBot.Core.Commands.Utility
             KaguyaEmbedBuilder embed;
 
             text = Regex.Replace(text, "[mM][yY]", "your");
-            text = text.ToLower();
-
+            text = text.Humanize(LetterCasing.Sentence);
+            
             if (time.TotalSeconds < 10)
             {
                 embed = new KaguyaEmbedBuilder